requirements.txt
Flask>=3.0
python-telegram-bot>=21.0
python-dateutil>=2.9
requests>=2.31
reportlab>=4.0
openpyxl>=3.1
